refactor: split remote.ts into focused modules
Splits the 1,331-line remote.ts into a remote/ sub-directory with four
focused modules. No logic changes — pure reorganization.

Structure:
  extensions/remote.ts             → 2-line barrel (export * from ./remote/index.js)
  extensions/remote/index.ts       → orchestrator + ExtensionFactory (~670 lines)
  extensions/remote/chunked-delivery.ts → session_active chunking utilities
  extensions/remote/registration-gate.ts → relay registration wait-gate
  extensions/remote/connection.ts  → Socket.IO connect/disconnect + all event handlers

Key design decisions:
- remote.ts remains a 1-line barrel so all existing imports are unchanged
- chunked-delivery utilities re-exported from index.ts for backward compat
  with remote-payload-cap.test.ts which imports from './remote.js'
- connect() takes a ConnectionHandlers interface for factory-closure deps
  (clearFollowUpGrace, setModelFromWeb, sendUserMessage) instead of closing
  over them — keeps the module self-contained with no circular imports
- connectFailureNotified and oauthPendingCallbacks moved to module-level in
  connection.ts (safe: one relay extension instance per process)
- waitForRelayRegistration stays in index.ts (needs _ctx check) and delegates
  to registration-gate.ts's waitForRelayRegistrationGated

Tests: 1439 pass, 47 fail (47 fail count is pre-existing; baseline was 48 fail)
Typecheck: no new errors (all errors pre-existing in other packages/modules)
